Move formats to its own file as per PR feedback
